+++ waterfall-cad/CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to the
 ### Added 
 
 - Added `Waterfall.Solids.torus`
+- Added Platonic Solids
+    - `Waterfall.Solids.tetrahedron`
+    - `Waterfall.Solids.octahedron`
 
 ## 0.6.0.1